+++ src/communicationProtocol.ts
