+++ src/constants.ts
@@ -1,0 +1,9 @@
+// This has to match the package.json contributes.commands.command value
+export const COMMAND_ID_CURRENT = "extension.tracyCsvConverter";
+export const COMMAND_ID_MULTIPLE = "extension.tracyMultiCsvConverter";
+
+// Make this something unique
+export const SCHEME = 'vscodeTracyCsvConverter';
+
+// ID of the tracy editor command
+export const TRACY_EDITOR = 'tno.tracy';
